Performance measures from effect estimation within all repeats of the simulation

Simulation for ?????? Repeats anzSim times the following steps

Determining the design matrix regarding the chosen design (Design, Number of cluster, time points, individuals) knowing Fidelity pattern and possible data loss

Sample data of the given study design and expected data loss using the package "samplingDataCRT"

Estimation of Effects using linear mixed model estimation

Usage

simulation(
  anzSim,
  type,
  sigma.1,
  sigma.2 = NULL,
  sigma.3,
  K,
  J,
  I,
  mu.0,
  theta,
  betas,
  X,
  X.A,
  B.cond = "0",
  C.cond = 0,
  D.cond = 0
)

Arguments

anzSim

Number of simulation repeats

type

Study design type = "cross-sec" for cross-sectional or "long" for longitudinal

sigma.1

Within variability or error variance (Hughes&Hussey sigma)

sigma.2

Between individual variance (only used when the individuals are followed over time: longitudinal study)

sigma.3

Between clusters variability (Hughes&Hussey tau)

K

Number of time points (measurement)

J

Number of Individuals (=individuals per cluster)

I

Number of cluster

mu.0

Baseline mean within the model specification

theta

Intervention effect

betas

Time trend could be included

X

Design matrix for building linear model of given ooptimal study

X.A

Design matrix for building linear model of sampling real data given real setting

B.cond

Condition for type of cluster loss: 0 = No cluster loss, 1 = Cluster missing at random, 2 = Cluster is missing at beginning, 3 = Cluster is missing at end of the study

C.cond

Condition for number of clusters which getting lost

D.cond

Condition for Number of individual loss (chosen randomly from all individuals within the study)

Value

Vector of Performance measures (see function performanceMeas):

Average of the estimates of the parameter of interest

Empirical standard error as an assessment of the estimation uncertainty

Bias = mean deviation of the estimates from the true value of the parameter of interest is an indicator of accuracy

Coverage of a confidence interval is a measurement that can be used to control the Type I error

Empirical power = the proportion of simulation samples in which the H0 of no effect is rejected at a significance level when H0 is false